Updated timeout waiting for security update checks
Updated the security update timeout (using waitmax) before the logic statement about whether the --security yum feature is enabled. The check was not completing on many servers due to the sheer number of security updates that are required to be processed by yum, and thus the plugin reported "-2" to the check. This gave false results in the check suggesting no updates were available when they are.